+++ src/cheatsheet/sections/scopes.py
@@ -1,0 +1,11 @@
+from ..get_list import get_lists
+
+
+def get_scopes():
+    return {
+        **get_lists(
+            ["scope_type", "selection_type", "subtoken_scope_type"],
+            {"argumentOrParameter": "Argument"},
+        ),
+        "<P>": "Paired delimiter",
+    }
